tc-framework-apps-python-flask-app-QAP-193/requirements.txt
Flask==2.2.5
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.1
selenium==4.8.3
pytest==7.4.4
opencv-python==4.10.0.84
numpy==2.1.2
img2pdf==0.5.1
pyhtml2pdf==0.0.7
webdriver-manager==4.0.2